pyproject.toml
[project]
name = "hw-scraper"
version = "0.1.0"
description = ""
authors = [
    {name = "ox2a-fuu",email = "ox2a.arsenault.jesse@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi (>=0.13.0,<0.14.0)",
    "lxml (>=6.0.1,<7.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "aiodns (>=3.5.0,<4.0.0)",
    "asyncio-throttle (>=1.0.2,<2.0.0)",
    "psutil (>=7.1.0,<8.0.0)"
]

[tool.poetry]
packages = [{include = "hw_scraper", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
